--- hw01-testsystem-csv/src/main/java/ru/otus/merets/Main.java
@@ -1,18 +1,0 @@
-package ru.otus.merets;
-
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.PropertySource;
-import ru.otus.merets.service.TestingServiceImpl;
-import ru.otus.merets.service.TestingService;
-
-@ComponentScan
-@PropertySource("classpath:application.properties")
-public class Main {
-
-    public static void main(String[] args) {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
-        TestingService testingService = context.getBean(TestingServiceImpl.class);
-        testingService.startTest();
-    }
-}
