--- hw01-testsystem-csv/src/main/java/ru/otus/merets/dao/CsvQuestionDaoParsingException.java
@@ -1,11 +1,0 @@
-package ru.otus.merets.dao;
-
-public class CsvQuestionDaoParsingException extends RuntimeException {
-    public CsvQuestionDaoParsingException(String message) {
-        super(message);
-    }
-
-    public CsvQuestionDaoParsingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
