--- hw01-testsystem-csv/src/main/java/ru/otus/merets/dao/QuestionDao.java
@@ -1,9 +1,0 @@
-package ru.otus.merets.dao;
-
-import ru.otus.merets.domain.Question;
-
-import java.util.List;
-
-public interface QuestionDao {
-    List<Question> getAllQuestions();
-}
