--- hw01-testsystem-csv/src/main/java/ru/otus/merets/domain/Answer.java
@@ -1,29 +1,0 @@
-package ru.otus.merets.domain;
-
-public class Answer {
-    private final String message;
-    private final Boolean isRight;
-    private final String id;
-
-    public Answer(String id, String message, Boolean isRight){
-        this.id = id;
-        this.message = message;
-        this.isRight = isRight;
-    }
-
-    public Boolean isCorrect(){
-        return isRight;
-    }
-
-    public String getId(){
-        return id;
-    }
-
-    @Override
-    public String toString() {
-        return " - " + id +
-                " " + message +
-                " (" + isRight +
-                ") ";
-    }
-}
