--- hw01-testsystem-csv/src/main/java/ru/otus/merets/service/IOService.java
@@ -1,6 +1,0 @@
-package ru.otus.merets.service;
-
-public interface IOService {
-    void printMessage(Object obj);
-    String getString();
-}
