--- hw01-testsystem-csv/src/main/java/ru/otus/merets/service/QuestionService.java
@@ -1,10 +1,0 @@
-package ru.otus.merets.service;
-
-import ru.otus.merets.domain.Question;
-
-import java.util.List;
-
-public interface QuestionService {
-    List<Question> getAllQuestions();
-    List<String> getCorrectAnswers(Question question);
-}
