--- hw01-testsystem-csv/src/main/java/ru/otus/merets/service/TestingService.java
@@ -1,8 +1,0 @@
-package ru.otus.merets.service;
-
-import ru.otus.merets.domain.Question;
-
-public interface TestingService {
-    void startTest();
-    boolean ask(Question question);
-}
