--- hw01-testsystem-csv/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-question.path=/questions.csv
-exam.passScore=3
