--- hw01-testsystem-csv/src/test/java/ru/otus/merets/TestContextConfig.java
@@ -1,35 +1,0 @@
-package ru.otus.merets;
-
-import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.PropertySource;
-import ru.otus.merets.service.*;
-
-@Configuration
-@PropertySource("classpath:application.properties")
-public class TestContextConfig {
-
-    @Bean
-    @Primary
-    QuestionService questionService(){
-        return Mockito.mock(QuestionServiceImpl.class);
-    }
-
-    @Bean
-    @Primary
-    IOService ioService(){
-        return Mockito.mock(IOServiceConsole.class);
-    }
-
-    @Value("${exam.passScore}")
-    private Integer passScore;
-
-    @Bean
-    @Primary
-    TestingService testingService(){
-        return new TestingServiceImpl(questionService(),ioService(), passScore);
-    }
-}
