--- hw01-testsystem-csv/src/test/resources/application.properties
@@ -1,2 +1,0 @@
-question.path=/questions-correct.csv
-exam.passScore=1
