+++ testsystem/src/main/java/ru/otus/merets/testsystem/config/ExamProperties.java
@@ -1,0 +1,36 @@
+package ru.otus.merets.testsystem.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.Locale;
+
+@ConfigurationProperties(prefix = "exam")
+public class ExamProperties {
+    private String path;
+    private Integer score;
+    private Locale locale;
+
+    public Locale getLocale() {
+        return locale;
+    }
+
+    public void setLocale(Locale locale) {
+        this.locale = locale;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+
+    public Integer getScore() {
+        return score;
+    }
+
+    public void setScore(Integer score) {
+        this.score = score;
+    }
+}
