+++ testsystem/src/main/java/ru/otus/merets/testsystem/config/Localization.java
@@ -1,0 +1,17 @@
+package ru.otus.merets.testsystem.config;
+
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+@Configuration
+public class Localization {
+    @Bean
+    public MessageSource messageSource() {
+        var ms = new ReloadableResourceBundleMessageSource();
+        ms.setBasename("classpath:/i18n/bundle");
+        ms.setDefaultEncoding("UTF-8");
+        return ms;
+    }
+}
