+++ testsystem/src/main/java/ru/otus/merets/testsystem/dao/CsvQuestionDaoParsingException.java
@@ -1,0 +1,11 @@
+package ru.otus.merets.testsystem.dao;
+
+public class CsvQuestionDaoParsingException extends RuntimeException {
+    public CsvQuestionDaoParsingException(String message) {
+        super(message);
+    }
+
+    public CsvQuestionDaoParsingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
