+++ testsystem/src/main/java/ru/otus/merets/testsystem/dao/QuestionDao.java
@@ -1,0 +1,9 @@
+package ru.otus.merets.testsystem.dao;
+
+import ru.otus.merets.testsystem.domain.Question;
+
+import java.util.List;
+
+public interface QuestionDao {
+    List<Question> getAllQuestions();
+}
