+++ testsystem/src/main/java/ru/otus/merets/testsystem/service/IOService.java
@@ -1,0 +1,8 @@
+package ru.otus.merets.testsystem.service;
+
+
+public interface IOService {
+    void printMessage(Object obj);
+
+    String getString();
+}
