+++ testsystem/src/main/java/ru/otus/merets/testsystem/service/QuestionService.java
@@ -1,0 +1,12 @@
+package ru.otus.merets.testsystem.service;
+
+
+import ru.otus.merets.testsystem.domain.Question;
+
+import java.util.List;
+
+public interface QuestionService {
+    List<Question> getAllQuestions();
+
+    List<String> getCorrectAnswers(Question question);
+}
