+++ testsystem/src/main/java/ru/otus/merets/testsystem/service/TestingService.java
@@ -1,0 +1,5 @@
+package ru.otus.merets.testsystem.service;
+
+public interface TestingService {
+    void startTest();
+}
