+++ testsystem/src/main/resources/i18n/bundle_en.properties
@@ -1,0 +1,5 @@
+messages.passed = Congratulations! You passed the exam.
+messages.failed = Unfortunately, you failed the exam. Try one more time.
+messages.ask_name = Please enter your name and press Enter:
+messages.greeting = After each question you should enter your answers (one or several). In case of many correct answers, please use space like a separator, e.g. '1 2 3'.
+messages.result ={0}, your score is {1} from {2} (necessary at least {3}).
