+++ testsystem/src/main/resources/i18n/bundle_ru_RU.properties
@@ -1,0 +1,5 @@
+messages.passed = Поздравляем! Вы успешно сдали экзамен.
+messages.failed = К сожалению, Вы не набрали необходимого балла. Попробуйте еще раз.
+messages.ask_name = Введите свое имя:
+messages.greeting = Вы должны вводить номер правильного ответа после каждого вопроса (один или несколько). Если в вопросе несколько правильных ответов - вводите их через пробел, пример: '1 2 3'.
+messages.result ={0}, Ваш балл: {1} (из {2}). Для прохождения необходимо {3}.
