+++ testsystem/src/test/resources/application.yml
@@ -1,0 +1,4 @@
+exam:
+  path: "questions.csv"
+  score: 1
+  locale: en
