+++ testsystem/src/test/resources/questions/ru_RU/questions.csv
@@ -1,0 +1,5 @@
+1,К методологии безопасности приложений относится,1,BSIMM,true,2,OWASP Top 10,false
+2,Можно ли использовать целочисленное представление вместо IP-адреса (например http://3258611659/ ),1,Нет,false,2,Да,true
+3,ТТП это,1,Транстихоокеанское партнерство,false,2,Тактики техники и процедуры,true,3,Тромботическая тромбоцитопеническая пурпура,false,4,Теорема о трех перепендикулярах,false
+4,Переполнение стека характерно для,1,C++,true,2,JavaScript,false
+5,Хакеры бывают,1,Черные,true,2,Белые,true,3,Синие,true,4,Зеленые,true
